# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Habilitations UFCV API
  version: 1.0.0
  description: API REST interne de gestion des habilitations UFCV.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /people/{id}:
    get:
      summary: Detail d'une personne
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail personne
    patch:
      summary: Modifier une personne
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Personne modifiee
...
  /documents/{id}/assign:
    post:
      summary: Assigner un document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Assignation creee
  /documents/{id}/sign:
    post:
      summary: Signer un document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Signature creee
...